+++ tea-coffee-core/functions.php
@@ -87,7 +87,7 @@ function cfe_splash_info( $wp_customize ) {
         'default' => 'Technology. Innovation. Us.'
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
-    cfe_splash-info-headline-control', array(
+    cfe-splash-info-headline-control', array(
         'label' => 'Headline',
         'section' => 'cfe-splash-info-section',
         'settings' => 'cfe-splash-info-headline'
@@ -97,7 +97,7 @@ function cfe_splash_info( $wp_customize ) {
         'default' => 'Community of '
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
-    cfe_splash-info-typing-prompt-control', array(
+    cfe-splash-info-typing-prompt-control', array(
         'label' => 'Typing prompt',
         'section' => 'cfe-splash-info-section',
         'settings' => 'cfe-splash-info-typing-prompt'
@@ -107,7 +107,7 @@ function cfe_splash_info( $wp_customize ) {
         'default' => '[ "Enthusiasts.", "Learning.", "Fun!" ]'
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
-    cfe_splash-info-typing-array-control', array(
+    cfe-splash-info-typing-array-control', array(
         'label' => 'Typing array',
         'section' => 'cfe-splash-info-section',
         'settings' => 'cfe-splash-info-typing-array'
@@ -117,21 +117,62 @@ function cfe_splash_info( $wp_customize ) {
         'default' => '500',
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
-    cfe_splash-info-typing-speed-control', array(
+    cfe-splash-info-typing-speed-control', array(
         'label' => 'Typing speed',
         'section' => 'cfe-splash-info-section',
         'settings' => 'cfe-splash-info-typing-speed',
     )));
 
     $wp_customize->add_setting('cfe-splash-general-info', array(
-        'default' => 'Drop by Rm 5370, SFU Surrey Campus every Monday for out work meetings'
+        'default' => 'Drop by Rm 5370, SFU Surrey Campus every Monday for our work meetings'
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
-    cfe_splash-general-info-control', array(
+    cfe-splash-general-info-control', array(
         'label' => 'Information',
         'section' => 'cfe-splash-info-section',
         'settings' => 'cfe-splash-general-info'
     )));
+}
+
+/**
+ * Add custom event section
+ */
+function cfe_event_info( $wp_customize ) {
+    $wp_customize->add_section('cfe_event_info-section', array(
+        'title' => 'Event Settings'
+    ));
+
+    $wp_customize->add_setting('cfe-event-year', array(
+        'default' => '2018'
+    ));
+    $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
+    cfe-event-year-control', array(
+        'label' => 'Headline',
+        'section' => 'cfe_event_info-section',
+        'settings' => 'cfe-event-year'
+    )));
+
+    $wp_customize->add_setting('cfe-event-datetime', array(
+        'default' => 'Saturday September 16th-17th 2017'
+    ));
+    $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
+    cfe-event-datetime-control', array(
+        'label' => 'Headline',
+        'section' => 'cfe_event_info-section',
+        'settings' => 'cfe-event-datetime'
+    )));
+
+    $wp_customize->add_setting('cfe-event-location', array(
+        'default' => 'SFU Surrey'
+    ));
+    $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
+    cfe-event-location-control', array(
+        'label' => 'Headline',
+        'section' => 'cfe_event_info-section',
+        'settings' => 'cfe-event-location'
+    )));
+
+    
 }
 
 /**
@@ -164,6 +205,7 @@ function WidgetLocation(){
 add_action( 'after_setup_theme', 'tea_coffee_setup' );
 add_action( 'wp_enqueue_scripts', 'wordpress_resources');
 add_action('customize_register', 'cfe_splash_info');
+add_action('customize_register', 'cfe_event_info');
 add_filter( 'excerpt_length', 'custom_excerpt_length', 999 );
 add_action('widgets_init', 'WidgetLocation');